+++ lib/r18gs/src/use-rgs.ts
@@ -1,46 +1,7 @@
 /* eslint-disable @typescript-eslint/non-nullable-type-assertion-style -- as ! operator is forbidden by eslint*/
-import { useSyncExternalStore } from "react";
+import { createHook, createSetter, createSubcriber, globalRGS } from "./utils";
 
-type Listener = () => void;
-type Subscriber = (l: Listener) => () => void;
-export type SetterArgType<T> = T | ((prevState: T) => T);
-export type SetStateAction<T> = (value: SetterArgType<T>) => void;
-
-/**
- * This is a hack to reduce lib size + readability + not encouraging direct access to globalThis
- */
-const [VALUE, LISTENERS, SETTER, SUBSCRIBER] = [0, 1, 2, 3];
-type RGS = [unknown, Listener[], SetStateAction<unknown>, Subscriber];
-
-declare global {
-	// eslint-disable-next-line no-var -- var required for global declaration.
-	var rgs: Record<string, RGS | undefined>;
-}
-
-const globalThisForBetterMinification = globalThis;
-globalThisForBetterMinification.rgs = {};
-const globalRGS = globalThisForBetterMinification.rgs;
-
-/** Initialize the named store when invoked for the first time. */
-function init<T>(key: string, value?: T) {
-	const listeners: Listener[] = [];
-	/** setter function to set the state. */
-	const setter: SetStateAction<T> = val => {
-		const rgs = globalRGS[key] as RGS;
-		rgs[VALUE] = val instanceof Function ? val(rgs[VALUE] as T) : val;
-		(rgs[LISTENERS] as Listener[]).forEach(listener => listener());
-	};
-	/** subscriber function to subscribe to the store. */
-	const subscriber: Subscriber = listener => {
-		const rgs = globalRGS[key] as RGS;
-		const listeners = rgs[LISTENERS] as Listener[];
-		listeners.push(listener);
-		return () => {
-			rgs[LISTENERS] = listeners.filter(l => l !== listener);
-		};
-	};
-	globalRGS[key] = [value, listeners, setter as SetStateAction<unknown>, subscriber];
-}
+import type { SetStateAction } from "./utils";
 
 /**
  * Use this hook similar to `useState` hook.
@@ -59,17 +20,9 @@ function init<T>(key: string, value?: T) {
  * @returns - A tuple (Ordered sequance of values) containing the state and a function to set the state.
  */
 export default function useRGS<T>(key: string, value?: T, serverValue?: T): [T, SetStateAction<T>] {
-	if (!globalRGS[key]) init(key, value);
+	/** Initialize the named store when invoked for the first time. */
+	if (!globalRGS[key])
+		globalRGS[key] = [value, serverValue, [], createSetter(key), createSubcriber(key)];
 
-	const rgs = globalRGS[key] as RGS;
-
-	/** Function to set the state. */
-	const setRGState = rgs[SETTER] as SetStateAction<T>;
-	/** Function to get snapshot of the state. */
-	const getSnap = () => (rgs[VALUE] ?? value) as T;
-	/** Function to get server snapshot. Returns server value is provided else the default value. */
-	const getServerSnap = () => (serverValue ?? rgs[VALUE] ?? value) as T;
-
-	const val = useSyncExternalStore<T>(rgs[SUBSCRIBER] as Subscriber, getSnap, getServerSnap);
-	return [val, setRGState];
+	return createHook(key);
 }